prorrogas

+++ src/Brasa/RecursoHumanoBundle/Controller/BaseEntidadExamenController.php
@@ -21,6 +21,7 @@ class BaseEntidadExamenController extends Controller
         $em = $this->getDoctrine()->getManager();
         $request = $this->getRequest(); // captura o recupera datos del formulario
         $paginator  = $this->get('knp_paginator');
+        $objMensaje = new \Brasa\GeneralBundle\MisClases\Mensajes();
         $form = $this->createFormBuilder() //
             ->add('BtnPdf', 'submit', array('label'  => 'PDF'))
             ->add('BtnExcel', 'submit', array('label'  => 'Excel'))
@@ -34,7 +35,13 @@ class BaseEntidadExamenController extends Controller
                 foreach ($arrSeleccionados AS $codigoEntidadExamenPk) {
                     $arEntidadExamen = new \Brasa\RecursoHumanoBundle\Entity\RhuEntidadExamen();
                     $arEntidadExamen = $em->getRepository('BrasaRecursoHumanoBundle:RhuEntidadExamen')->find($codigoEntidadExamenPk);
-                    $em->remove($arEntidadExamen);
+                    $arEntidadExamenDetalle = $em->getRepository('BrasaRecursoHumanoBundle:RhuExamenListaPrecio')->findOneBy(array('codigoEntidadExamenFk' => $codigoEntidadExamenPk));
+                    if ($arEntidadExamenDetalle == null){
+                        $em->remove($arEntidadExamen);
+                    } else {
+                        $objMensaje->Mensaje("error", "No se puede eliminar el registro " . $codigoEntidadExamenPk .", tiene detalles asociados", $this);
+                    }
+                    
                     $em->flush();
                 }
             }
